pytorch-flask-deploy/requirements.txt
Flask==1.1.1
gevent==1.4.0
numpy==1.19.5
Pillow==8.3.2
Werkzeug==2.0.1
torch==1.12.0
splitfolders==0.5.1
torchvision==0.11.2+cu113